% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerLoad.R
\name{playerLoad}
\alias{playerLoad}
\title{A Player Load Function}
\usage{
playerLoad(pageUrl, removeRow = TRUE)
}
\arguments{
\item{removeRow}{Removes top row from the source table if it contains two headers. Default to TRUE.}

\item{pageURL}{Hokej.cz page url with player stats table.}
}
\description{
This function allows you to download player data from Hokej.cz
}
\examples{
pageUrl <- "http://hokej.cz/tipsport-extraliga/stats-center/player-analytic?season=2017&competition=6026&yearFrom=&stranger=0&state=&stats-section=shots&stats-all=1"
shots <- playerLoad(pageUrl)
}
\keyword{player}
