% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamLoad.R
\name{teamLoad}
\alias{teamLoad}
\title{A Team Load Function}
\usage{
teamLoad(pageUrl, removeRow = TRUE)
}
\arguments{
\item{removeRow}{Removes top row from the source table if it contains two headers. Default to TRUE.}

\item{pageURL}{Hokej.cz page url with team stats table.}
}
\description{
This function allows you to download team data from Hokej.cz
}
\examples{
pageUrl <- "http://hokej.cz/tipsport-extraliga/stats-center/team-analytic-stats?season=2017&competition=6026&stats-section=corsi"
corsiTeam <- teamLoad(pageUrl,removeRow = F)
}
\keyword{player}
